Dynamic Extension Settings & User Overridden Person mapping (#130)
* Refactored how people are mapped to ComicInfo roles to be more generic and flexible. Changed a few mappings to be more consistent with expectations.

* Refactored how people are mapped to ComicInfo roles to be more generic and flexible. Changed a few mappings to be more consistent with expectations.

* Moved the mapping into the class itself, which will avoid some memory pressure and enable future loading from settings.

* Refactored the code to use DefaultSettings which is class based settings. This is the first step, next step is to refactor to use Setttings() which will allow persistence.

* Refactored the code to load from settings.ini and be available via Settings singleton.

* Refactored execute_gui to use new Settings

* Refactored everything but Setting GUI code to use Setting()

* Made it so scraper settings can be loaded at runtime.

* Refactored all the widgets to be in their own submodule to make the code easier to maintain. Refactored (broken) how provider settings are being handled again to hopefully streamline and keep extensions settings self-contained.

* Updated the gitignore to properly cover python and JetBrains IDEs.

* Refactored the widgets to correctly use the Widget class Promidius wrote and thus made that Widget have a custom name to help differentiate.

* Updated AniList to have it's own settings for mapping

* Added some basic unit test for settings

* Refactored the Settings to have default settings that will be populated on initialization, if the settings.ini is empty. Thus SettingsWidgetManager needed to be refactored to have mapping of Setting keys to Controls and generic code to perform the mapping.

* Updated GA to Python 3.11 and updated .idea files to be more generic.

* Fixed a circular import issue. Refactored MetadataSources to use string enums for keys to reduce duplication of strings. When setting a setting, Settings will now create a new section if needed.

* MetadataSources will now try to set their defaults to Settings on load, but will not overwrite custom settings.
