+++ .github/workflows/Run_Tests.yml
@@ -15,7 +15,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: [ '3.10' ]
+        python-version: [ '3.11' ]
     name: Python ${{ matrix.python-version }} sample
     steps:
         -   uses: actions/checkout@v3
