+++ .idea/MangaManagerV2.iml
@@ -7,17 +7,10 @@
       <excludeFolder url="file://$MODULE_DIR$/build" />
       <excludeFolder url="file://$MODULE_DIR$/.pytest_cache" />
       <excludeFolder url="file://$MODULE_DIR$/dist" />
+      <excludeFolder url="file://$MODULE_DIR$/venv" />
     </content>
-    <orderEntry type="jdk" jdkName="Python 3.11 (MangaManagerV2_3.11)" jdkType="Python SDK" />
+    <orderEntry type="jdk" jdkName="Python 3.10 (Manga-Manager)" jdkType="Python SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
     <orderEntry type="library" name="jquery-3.6.0" level="application" />
-  </component>
-  <component name="TemplatesService">
-    <option name="TEMPLATE_FOLDERS">
-      <list>
-        <option value="$MODULE_DIR$/MangaManager/src/Common/Templates" />
-        <option value="$MODULE_DIR$/TestWeb/Templates" />
-      </list>
-    </option>
   </component>
 </module>
