+++ .idea/runConfigurations/Unittests.xml
@@ -3,10 +3,9 @@
     <module name="MangaManagerV2" />
     <option name="INTERPRETER_OPTIONS" value="" />
     <option name="PARENT_ENVS" value="true" />
-    <option name="SDK_HOME" value="C:\Users\galla\.virtualenvs\MangaManagerV2\Scripts\python.exe" />
-    <option name="SDK_NAME" value="Python 3.11 (MangaManagerV2)" />
+    <option name="SDK_HOME" value="" />
     <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$/MangaManager" />
-    <option name="IS_MODULE_SDK" value="false" />
+    <option name="IS_MODULE_SDK" value="true" />
     <option name="ADD_CONTENT_ROOTS" value="false" />
     <option name="ADD_SOURCE_ROOTS" value="false" />
     <EXTENSION ID="PythonCoverageRunConfigurationExtension" runner="coverage.py" />
