+++ MangaManager/Extensions/CoverDownloader/CoverDownloader.py
@@ -1,10 +1,6 @@
 from tkinter import Label, Frame, Entry
 
-from src import settings_class
 from src.DynamicLibController.models.CoverSourceInterface import Cover
-
-settings = settings_class.get_setting("main")
-covers_folder_path = settings.covers_folder_path
 
 
 def get_cover_from_source_dummy() -> list[Cover]:
