+++ MangaManager/ExternalSources/MetadataSources/__init__.py
@@ -1,0 +1,3 @@
+from .MangaUpdates import MangaUpdates
+from .AniList import AniList
+from .MetadataSourceFactory import ScraperFactory
