+++ MangaManager/ExternalSources/__init__.py
@@ -1,0 +1,1 @@
+from .MetadataSources.MetadataSourceFactory import ScraperFactory
