+++ MangaManager/src/MetadataManager/MetadataManagerGUI.py
@@ -11,6 +11,9 @@ from pkg_resources import resource_filename
 
 from src.Common.utils import get_platform, open_folder
 from src.MetadataManager import comicinfo
+from src.MetadataManager.GUI.ControlManager import ControlManager
+from src.Settings.SettingHeading import SettingHeading
+from src.Settings.Settings import Settings
 
 if get_platform() == "linux":
     from src.MetadataManager.GUI.FileChooserWindow import askopenfiles,askdirectory
@@ -19,14 +22,12 @@ else:
 from _tkinter import TclError
 
 from src.Common.loadedcomicinfo import LoadedComicInfo
-from src.MetadataManager.GUI.widgets import ComboBoxWidget, OptionMenuWidget, WidgetManager, ButtonWidget, ControlManager, HyperlinkLabel
-from src.MetadataManager.GUI.settings import SettingsWidgetManager
+from src.MetadataManager.GUI.widgets import ComboBoxWidget, OptionMenuWidget, WidgetManager, ButtonWidget, \
+    HyperlinkLabelWidget
+from src.MetadataManager.GUI.SettingsWidgetManager import SettingsWidgetManager
 from src.MetadataManager.MetadataManagerLib import MetadataManagerLib
 
-from src import settings_class
 from src.__version__ import __version__
-
-main_settings = settings_class.get_setting("main")
 
 
 class GUIApp(Tk, MetadataManagerLib):
@@ -113,7 +114,7 @@ class GUIApp(Tk, MetadataManagerLib):
         # These are some tricks to make it easier to select files.
         # Saves last opened folder to not have to browse to it again
         if not self.last_folder:
-            initial_dir = main_settings.library_path.value
+            initial_dir = Settings().get(SettingHeading.Main, 'library_path')
         else:
             initial_dir = self.last_folder
         self.log.debug("Selecting files")
@@ -152,14 +153,14 @@ class GUIApp(Tk, MetadataManagerLib):
         # These are some tricks to make it easier to select files.
         # Saves last opened folder to not have to browse to it again
         if not self.last_folder:
-            initial_dir = main_settings.library_path.value
+            initial_dir = Settings().get(SettingHeading.Main, 'library_path')
         else:
             initial_dir = self.last_folder
         self.log.debug("Selecting files")
         # Open select files dialog
 
         folder_path = askdirectory(initialdir=initial_dir)
-        self.selected_files_path = glob.glob(root_dir=folder_path,pathname=os.path.join(folder_path,"**/*.cbz"),recursive=True)
+        self.selected_files_path = glob.glob(root_dir=folder_path,pathname=os.path.join(folder_path, "**/*.cbz"),recursive=True)
         # TODO: Auto select recursive or not
         # self.selected_files_path = [str(Path(folder_path, file)) for file in os.listdir(folder_path) if file.endswith(".cbz")]
 
@@ -180,11 +181,11 @@ class GUIApp(Tk, MetadataManagerLib):
         top_level = tkinter.Toplevel(self)
         frame = Frame(top_level)
         frame.pack(pady=30, padx=30,fill="both")
-        HyperlinkLabel(frame, "Github repo:",url_text="Go to Github rework main page",url="https://github.com/MangaManagerORG/Manga-Manager/tree/rework/master").pack(fill="x", expand=True, side="top", anchor="center")
-        HyperlinkLabel(frame, "Get support:", url_text="Join MangaManager channel in Kavita discord",url="https://discord.gg/kavita-821879810934439936").pack(fill="x", expand=True, side="top", anchor="center")
-        HyperlinkLabel(frame, "Report issue in GitHub",url_text="Create GitHub Issue", url="https://github.com/MangaManagerORG/Manga-Manager/issues/new?assignees=ThePromidius&labels=Rework+Issue&template=rework_issue.md&title=%5BRework+Issue%5D").pack(
+        HyperlinkLabelWidget(frame, "Github repo:", url_text="Go to Github rework main page", url="https://github.com/MangaManagerORG/Manga-Manager/tree/rework/master").pack(fill="x", expand=True, side="top", anchor="center")
+        HyperlinkLabelWidget(frame, "Get support:", url_text="Join MangaManager channel in Kavita discord", url="https://discord.gg/kavita-821879810934439936").pack(fill="x", expand=True, side="top", anchor="center")
+        HyperlinkLabelWidget(frame, "Report issue in GitHub",url_text="Create GitHub Issue", url="https://github.com/MangaManagerORG/Manga-Manager/issues/new?assignees=ThePromidius&labels=Rework+Issue&template=rework_issue.md&title=%5BRework+Issue%5D").pack(
             fill="x", expand=True, side="top", anchor="center")
-        HyperlinkLabel(frame, "Donate in Ko-fi",
+        HyperlinkLabelWidget(frame, "Donate in Ko-fi",
                        "https://ko-fi.com/thepromidius").pack(fill="x", expand=True, side="top", anchor="center")
         tkinter.Label(frame, text="", font=("Helvetica", 12), justify="left").pack(fill="x", expand=True, side="top", anchor="center")
 
@@ -299,7 +300,7 @@ class GUIApp(Tk, MetadataManagerLib):
         self.widget_mngr.clean_widgets()
         if loaded_cinfo_list is None:
             loaded_cinfo_list = self.selected_items
-        if main_settings.cache_cover_images:
+        if Settings().get(SettingHeading.Main, 'cache_cover_images'):
             self.image_cover_frame.update_cover_image(loaded_cinfo_list)
 
         # Iterate all cinfo tags. Should there be any values that are not equal. Show "different values selected"
