+++ MangaManager/src/MetadataManager/MetadataManagerLib.py
@@ -4,19 +4,21 @@ import abc
 import logging
 from abc import ABC
 
-from src import settings_class, sources_factory
+from ExternalSources import ScraperFactory
+
+from src import sources_factory
 from src.Common.errors import EditedCinfoNotSet, MangaNotFoundError
 from src.Common.errors import NoComicInfoLoaded, CorruptedComicInfo, BadZipFile
 from src.Common.loadedcomicinfo import LoadedComicInfo
 from src.Common.terminalcolors import TerminalColors as TerCol
 from src.MetadataManager import comicinfo
 from src.MetadataManager.comicinfo import ComicInfo
+from src.Settings.SettingHeading import SettingHeading
+from src.Settings.Settings import Settings
 
 AniList = [source for source in sources_factory.get("MetadataSources") if source.name == "AniList"]
 
 logger = logging.getLogger("MetaManager.Core")
-settings = settings_class.get_setting("main")
-source_settings = settings_class.get_setting("ExternalSources")
 
 
 class _IMetadataManagerLib(abc.ABC):
@@ -175,7 +177,7 @@ class MetadataManagerLib(_IMetadataManagerLib, ABC):
         for file_path in self.selected_files_path:
             try:
                 loaded_cinfo = LoadedComicInfo(path=file_path)
-                if settings.cache_cover_images and not self.is_cli:
+                if Settings().get(SettingHeading.Main, 'cache_cover_images') and not self.is_cli:
                     loaded_cinfo.load_all()
                 else:
                     loaded_cinfo.load_metadata()
@@ -209,15 +211,13 @@ class MetadataManagerLib(_IMetadataManagerLib, ABC):
         # loaded_cinfo.cinfo_object.export(export, 0)
         # print(export.getvalue())
 
-    def fetch_online(self,series_name):
-
-        selected_source = [source for source in sources_factory["MetadataSources"] if source.name == source_settings.default_metadata_source.value]
+    def fetch_online(self, series_name):
+        selected_source = ScraperFactory().get_scraper(Settings().get(SettingHeading.ExternalSources, 'default_metadata_source'))
         if not selected_source:
             raise Exception("Unhandled exception. Metadata sources are not loaded or there's a bug in it."
                             "Raise an issue if this happens.")
-        source = selected_source[0]
         try:
-            return source.get_cinfo(series_name)
+            return selected_source.get_cinfo(series_name)
         except MangaNotFoundError as e:
             logger.exception(str(e))
             self.on_manga_not_found(e, series_name)
