+++ MangaManager/src/Settings/SettingControlType.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class SettingControlType(Enum):
+    Bool = 0,
+    Text = 1,
+    Options = 2
