+++ MangaManager/src/Settings/SettingHeading.py
@@ -1,0 +1,7 @@
+from enum import StrEnum
+
+
+class SettingHeading(StrEnum):
+    Main = "Main",
+    WebpConverter = "Webp Converter",
+    ExternalSources = "External Sources",
