+++ MangaManager/src/Settings/__init__.py
@@ -1,0 +1,5 @@
+from .SettingControlType import SettingControlType
+from .SettingHeading import SettingHeading
+from .SettingControl import SettingControl
+from .SettingSection import SettingSection
+from .Settings import Settings
